+++ src/frontend-cdoc-commons/components/cdoc-albumpage/cdoc-albumpage.component.ts
@@ -460,7 +460,7 @@ export abstract class CommonDocAlbumpageComponent <R extends CommonDocRecord, F 
         this.pageUtils.setGlobalStyle('.hide-on-fullpage { display: none; } ' +
             '.show-on-fullpage-block { display: block; } ' +
             'body { background: #130b0b; } ' +
-            '.image-content-container {background: #130b0b !IMPORTANT; border: none !IMPORTANT;} ', 'fullPageStyle');
+            '.image-content-container {background: #130b0b !important; border: none !important;} ', 'fullPageStyle');
     }
 
     protected setEditPageLayoutAndStyles(): void {